report/requirements.txt
xhtml2pdf
python-dotenv
boto3
psycopg2